+++ js/util.js
@@ -1,0 +1,2 @@
+const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+export default {getRandomInt};
